+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Deleting a token will delete its tooltip.
 - Tooltip size can be adjusted via user settings; each user can customize this individually.
 - Improved visuals for Tooltip Adversary, Companion and characters.
+- DowntimeUI: Automated Beastbound
 
 # 0.2.2
 - Tooltip for Adversaries